third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/letter-spacing-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'letter-spacing' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'letter-spacing' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'letter-spacing' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'letter-spacing' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
Harness: the test ran to completion.
